app/requirements.txt
cassandra-driver
pyarrow
fastparquet
pathvalidate
pyspark
tqdm
venv-pack
cqlsh